Added unit tests

+++ hotel-ws/src/main/java/edu/sjsu/codebrigade/hotelws/dto/Hotel.java
@@ -2,15 +2,15 @@ package edu.sjsu.codebrigade.hotelws.dto;
 
 public class Hotel {
 
-    private long id;
+    private int id;
     private String name;
 
-    public Hotel(long id, String name) {
+    public Hotel(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
